requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.16

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Validation
pydantic==2.10.4
pydantic-settings==2.6.1
email-validator==2.2.0

# CORS
fastapi-cors==0.0.6

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1

# Development
black==24.10.0
flake8==7.1.1
isort==5.13.2

# Utilities
python-dateutil==2.9.0
pytz==2025.1
python-dotenv==1.0.1

# Supabase
supabase==2.17.0